DeconViews constructor: actually use the given ExecutorService, instead of creating one and ignoring the argument.
While the service could be obtained with getExecutorService,
code that provided a valid service did not used this method
to retrieve the ExecutorService created within the constructor,
which was therefore left dangling without shutdown.

The lack of shutdown results in a native memory leak, eventually
preventing the creation of new threads.